Multi: split models into multiple files
each module now has its own models.py, which makes it much more
managable.
