+++ app/calendar/routes.py
@@ -1,11 +1,11 @@
 from app import db
-from app.helpers import page_title, admin_required, stretch_color
-from app.models import CalendarSetting, Epoch, Month, Day, Moon
 from app.calendar import bp
 from app.calendar.forms import EpochForm, MonthForm, DayForm, MoonForm
+from app.calendar.models import CalendarSetting, Epoch, Month, Day, Moon
 from app.calendar.helpers import get_next_epoch_order, get_next_month_order, get_next_day_order, calendar_sanity_check, gen_calendar_preview_data, gen_calendar_stats, get_years_in_epoch, get_epochs, gen_epoch_choices, gen_month_choices, gen_day_choices
 from app.event.forms import EventForm
 from app.event.helpers import get_event_categories
+from app.helpers import page_title, admin_required, stretch_color
 from flask import render_template, flash, redirect, url_for, request
 from flask_login import login_required, current_user
