+++ app/character/routes.py
@@ -2,8 +2,9 @@ from app import db
 from app.character import bp
 from app.character.forms import CreateCharacterForm, EditCharacterForm, JournalForm
 from app.character.helpers import gen_session_choices
+from app.character.models import Character, Journal
 from app.helpers import page_title, flash_no_permission
-from app.models import Character, Party, Journal
+from app.party.models import Party
 from datetime import datetime
 from flask import render_template, flash, redirect, url_for, jsonify, request
 from flask_login import current_user, login_required
