+++ app/event/routes.py
@@ -1,11 +1,12 @@
 from app import db
-from app.helpers import page_title, flash_no_permission, stretch_color
-from app.models import EventSetting, Event, EventCategory, Epoch, Moon
-from app.user.models import User, Role
+from app.calendar.models import Epoch, Moon
+from app.calendar.helpers import gen_calendar_stats, gen_epoch_choices, gen_month_choices, gen_day_choices
 from app.event import bp
 from app.event.forms import SettingsForm, EventForm, CategoryForm
 from app.event.helpers import event_admin_required, update_timestamp, get_events, gen_event_category_choices, get_events_by_category
-from app.calendar.helpers import gen_calendar_stats, gen_epoch_choices, gen_month_choices, gen_day_choices
+from app.event.models import EventSetting, Event, EventCategory
+from app.helpers import page_title, flash_no_permission, stretch_color
+from app.user.models import User, Role
 from flask import render_template, flash, redirect, url_for, request, jsonify
 from flask_login import login_required, current_user
 from sqlalchemy import not_, and_, or_
