+++ app/map/routes.py
@@ -3,8 +3,9 @@ from app.helpers import page_title, flash_no_permission, admin_required
 from app.map import bp
 from app.map.forms import MapNodeTypeCreateForm, MapNodeTypeEditForm, MapSettingsForm, MapNodeForm, MapForm
 from app.map.helpers import map_admin_required, map_node_filename, gen_node_type_choices, get_visible_nodes, map_changed, gen_submap_choices
-from app.models import Map, MapNodeType, MapSetting, MapNode, WikiEntry
+from app.map.models import Map, MapNodeType, MapSetting, MapNode
 from app.wiki.helpers import gen_wiki_entry_choices
+from app.wiki.models import WikiEntry
 from datetime import datetime
 from flask import render_template, flash, redirect, url_for, request, jsonify, send_from_directory, current_app
 from flask_login import current_user, login_required
