+++ app/media/routes.py
@@ -1,10 +1,10 @@
 from app import db
 from app.helpers import page_title, flash_no_permission
-from app.models import MediaSetting, MediaItem, MediaCategory
-from app.user.models import User, Role
 from app.media import bp
 from app.media.forms import SettingsForm, MediaItemCreateForm, MediaItemEditForm, CategoryForm
 from app.media.helpers import media_admin_required, get_media, gen_media_category_choices, media_filename
+from app.media.models import MediaSetting, MediaItem, MediaCategory
+from app.user.models import User, Role
 from flask import render_template, flash, redirect, url_for, request, jsonify, send_from_directory, current_app
 from flask_login import login_required, current_user
 from sqlalchemy import not_, and_, or_
