+++ app/user/routes.py
@@ -48,6 +48,7 @@ def edit(username):
                 new_user_roles = Role.query.filter(Role.id.in_(form.roles.data)).all()
 
                 admin_role = Role.query.get(1)
+                admin_role.color = "cc9900"
 
                 if username == current_user.username and current_user.has_admin_role() and admin_role not in new_user_roles:
                     new_user_roles.append(admin_role)
